catalog item update & dropzone
