+++ client/src/lib/utils.ts
@@ -1,18 +1,6 @@
 import { type ClassValue, clsx } from "clsx";
 import { twMerge } from "tailwind-merge";
-import { z, ZodTypeAny } from "zod";
 
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs));
 }
-
-export const zodInputStringPipe = (zodPipe: ZodTypeAny) =>
-  z
-    .string()
-    .transform((value) => (value === "" ? null : value))
-    .nullable()
-    .refine((value) => value === null || !isNaN(Number(value)), {
-      message: "Значение должно быть числом",
-    })
-    .transform((value) => (value === null ? 0 : Number(value)))
-    .pipe(zodPipe);
